102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,18 +8,18 @@
 int main(void)
 {
 int i = 0;
-int j = 1, k = 2;
+long j = 1, k = 2;
 while (i < 50)
 {
 if (i == 0)
-printf("%d", j);
+printf("%lf", j);
 else if (i == 1)
-printf(", %d", k);
+printf(", %lf", k);
 else
 {
 k += j;
 j = k - j;
-printf(", %d", k);
+printf(", %lf", k);
 }
 ++i;
 }
